move common_models.py
