+++ torchvision_models/lane_detection/__init__.py
@@ -1,2 +1,1 @@
 from .lstr import *
-from .common_models import *
